# handlers/auth.go
package handlers

import (
	"context"
	"crypto/sha256"
	"net/http"
	"recipes-api/models"
	"time"

	"github.com/gin-contrib/sessions"
	"github.com/gin-gonic/gin"
	"github.com/golang-jwt/jwt/v5"
	"github.com/rs/xid"
	"go.mongodb.org/mongo-driver/bson"
	"go.mongodb.org/mongo-driver/mongo"
)

type Claims struct {
	Username string `json:"username"`
	jwt.RegisteredClaims
}

type JWTOutput struct {
	Token   string    `json:"token"`
	Expires time.Time `json:"expires"`
}

type AuthHandler struct {
	collection *mongo.Collection
	ctx        context.Context
}

func NewAuthHandler(ctx context.Context, collection *mongo.Collection) *AuthHandler {
	return &AuthHandler{
		collection: collection,
		ctx:        ctx,
	}
}

// SignOutHandler godoc
//
//	@Summary		Signout
//	@Description	Signout user
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Param			recipe	body		models.User	true "some"
//	@Success		200		{object}	string
//	@Router			/signout [post]
func (handler *AuthHandler) SignOutHandler(c *gin.Context) {
	session := sessions.Default(c)
	session.Clear()
	session.Save()
	c.JSON(http.StatusOK, gin.H{"message": "Signed out..."})
}

// SignInHandler godoc
//
//	@Summary		Signin
//	@Description	Sign in user
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Param			recipe	body		models.User	true "comment"
//	@Success		200		{object}	string
//	@Router			/signin [post]
func (handler *AuthHandler) SignInHandler(c *gin.Context) {
	var user models.User
	if err := c.ShouldBindJSON(&user); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	h := sha256.New()
	cur := handler.collection.FindOne(handler.ctx, bson.M{
		"username": user.Username,
		"password": string(h.Sum([]byte(user.Password))),
	})

	if cur.Err() != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid username or password"})
		return
	}

	sessionToken := xid.New().String()
	session := sessions.Default(c)
	session.Set("username", user.Username)
	session.Set("token", sessionToken)
	session.Save()

	c.JSON(http.StatusOK, gin.H{"message": "User signed in"})
}

// RefreshHandler godoc
//
//	@Summary		Refresh session
//	@Description	refresh session for user
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Param			recipe	body		models.User	true "comment"
//	@Success		200		{object}	string
//	@Router			/refresh [post]
func (handler *AuthHandler) RefreshHandler(c *gin.Context) {

	session := sessions.Default(c)
	sessionToken := session.Get("token")
	sessionUser := session.Get("username")

	if sessionToken == nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid session cookie"})
		return
	}

	sessionToken = xid.New().String()
	session.Set("username", sessionUser.(string))
	session.Set("token", sessionToken)
	session.Save()

	c.JSON(http.StatusOK, gin.H{"message": "New session issued"})
}

func (handler *AuthHandler) AuthMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {
		session := sessions.Default(c)
		sessionToken := session.Get("token")

		if sessionToken == nil {
			c.JSON(http.StatusForbidden, gin.H{
				"message": "Not logged",
			})
			c.Abort()
		}
		c.Next()
	}
}